+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.6.37 
+- Table name inference
+- Sql formatter
+- Record History
+
 ## 1.6.36 (2020/3/19)
 - Fix many mysql connection error.